requirements.txt
# DRIFT: Diffusion Reverse Inversion Fingerprint Tracking
# Core dependencies (shared with DRIFT_new)

# PyTorch ecosystem
torch>=2.0.0
torchvision>=0.15.0
torchaudio>=2.0.0

# Scientific computing
numpy>=1.24.0
scipy>=1.10.0
scikit-learn>=1.2.0

# Image processing
Pillow>=9.5.0
opencv-python>=4.7.0
imageio>=2.28.0
scikit-image>=0.20.0

# Data management
pandas>=2.0.0
PyYAML>=6.0

# Visualization
matplotlib>=3.7.0
seaborn>=0.12.0

# Logging and experiment tracking
tensorboard>=2.13.0
tqdm>=4.65.0

# Optional: for t-SNE
umap-learn>=0.5.3

# Jupyter (optional, for notebooks)
jupyter>=1.0.0
ipywidgets>=8.0.0
